requirements.txt
langchain>=0.1.16
langchain-google-genai>=1.0.0
google-generativeai>=0.5.0
tenacity>=8.2.3
python-dotenv>=1.0.1
gdpc
numpy